feat: 修改模型删除方案

+++ modules/cache.py
@@ -101,7 +101,7 @@ def cached_data_for_file(subsection, title, filename, func):
     If the data generation fails, None is returned to indicate the failure. Otherwise, the generated
     or cached data is returned as a dictionary.
     """
-    loadHuggingfaceModel(filename)
+    downloaded = loadHuggingfaceModel(filename)
     existing_cache = cache(subsection)
     ondisk_mtime = os.path.getmtime(filename)
 
@@ -120,5 +120,6 @@ def cached_data_for_file(subsection, title, filename, func):
         existing_cache[title] = entry
 
         dump_cache()
-    os.remove(filename)
+    if downloaded:
+        os.remove(filename)
     return entry['value']
